feat: Move most of the `Config` impl into the trait.
This PR moves the implementations of `headers`, `url`, and
`query` from the `OpenAIConfig` struct to the `Config` trait,
so other config implementations can override just the attribute
methods without having to re-implement the common material
(and take a direct dependency on `reqwest` in the process).

Note that this PR also adds `org_id` as a trait method, because
`headers` needs it. For backwards-compatibility, it
implements a default of empty string rather than being required.