feat: string-error feature allows deserialization into custom error format
the error format of different providers sometimes diverges from OpenAI.
the string-error feature does not try to parse errors into the OpenAI
format, but leaves it up to the user. when enabled, the  is
a wrapper around the original error message. This also enables cases
where the error message is not even valid json